feat(auth): :passport_control: complete auth validation
Login and registration validation fields completed.
