+++ src/controllers/auth.controllers.js
@@ -1,18 +1,8 @@
 import { response } from 'express'; // Se importa response y se lo agrega en el parametro res para obtener el autocompletado de las propiedades de res
-import { validationResult } from 'express-validator';
 
 export const loginUser = (req, res = response) => {
   try {
     const { email, password } = req.body;
-
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        ok: false,
-        errors: errors.mapped(),
-      });
-    }
 
     res.json({
       ok: true,
@@ -32,15 +22,6 @@ export const loginUser = (req, res = response) => {
 export const createUser = (req, res = response) => {
   try {
     const { name, email, password } = req.body;
-
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        ok: false,
-        errors: errors.mapped(),
-      });
-    }
 
     res.status(201).json({
       ok: true,
