+++ src/middlewares/validate-errors.js
@@ -1,0 +1,17 @@
+import { response } from 'express';
+import { validationResult } from 'express-validator';
+
+const validateErrors = (req, res = response, next) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).json({
+      ok: false,
+      errors: errors.mapped(),
+    });
+  }
+
+  next();
+};
+
+export default validateErrors;
