+++ src/routes/auth.routes.js
@@ -9,56 +9,17 @@ import {
   loginUser,
   renewToken,
 } from '../controllers/auth.controllers';
-import { check } from 'express-validator';
+import validateLoginFields from '../middlewares/auth/validate-fields/login-fields';
+import validateErrors from '../middlewares/validate-errors';
+import validateRegisterFields from '../middlewares/auth/validate-fields/register-fields';
 
 const router = Router();
 
 // Login Route
-router
-  .route('/')
-  .post(
-    [
-      check('email')
-        .notEmpty()
-        .withMessage('email is required')
-        .isEmail()
-        .withMessage('email must be valid')
-        .isLength({ min: 2, max: 50 })
-        .withMessage('email must be between 2 and 50 characters'),
-      check('password')
-        .notEmpty()
-        .withMessage('password is required')
-        .isLength({ min: 8, max: 12 })
-        .withMessage('password must be between 8 and 12 characters'),
-    ],
-    loginUser
-  );
+router.route('/').post([validateLoginFields, validateErrors], loginUser);
 
 // Register Route
-router
-  .route('/new')
-  .post(
-    [
-      check('name')
-        .notEmpty()
-        .withMessage('name is required')
-        .isLength({ min: 2, max: 10 })
-        .withMessage('name must be between 2 and 10 characters'),
-      check('email')
-        .notEmpty()
-        .withMessage('email is required')
-        .isEmail()
-        .withMessage('email must be valid')
-        .isLength({ min: 2, max: 50 })
-        .withMessage('email must be between 2 and 50 characters'),
-      check('password')
-        .notEmpty()
-        .withMessage('password is required')
-        .isLength({ min: 8, max: 12 })
-        .withMessage('password must be between 8 and 12 characters'),
-    ],
-    createUser
-  );
+router.route('/new').post([validateRegisterFields, validateErrors], createUser);
 
 // Renew Token Route
 router.route('/renew').get(renewToken);
